fix: make `ErrorKind` visible outside the library
While the struct `ErrorKind` is defined as `pub` inside the module
`error`, the module itself is private.
In contrast to the struct `Error`, `ErrorKind` is not exported by a
`pub use` in `lib.rs`, effectively making it invisible outside the
library and impeding fine-grained error handling.

This change fixes this by exporting `ErrorKind` alongside `Error`.